refactor(commands): extract agent-window / diagnostics / thinking pickers
Move the remaining command helpers out of extension.ts into commands/:
- commands/agentsWindow.ts — ensure/revert agent-window support + warm
  model-picker metadata
- commands/diagnostics.ts — showModelPickerDiagnostics
- commands/thinkingPicker.ts — showThinkingEffortPicker

extension.ts is now ~440 lines and holds only activate() wiring + deactivate
(no business logic). Behavior-preserving; compile + 291 tests + lint green.
