+++ src/commands/diagnostics.ts
@@ -1,0 +1,41 @@
+import * as vscode from "vscode";
+import { CONFIG_SECTION, SETTING_AGENTS_WINDOW } from "../config";
+import { AGENT_GO_VENDOR, AGENT_ZEN_VENDOR, GO_VENDOR, ZEN_VENDOR } from "../providerTypes";
+
+/** Dump every visible model + configuration schema into a Markdown doc. */
+export async function showModelPickerDiagnostics(): Promise<void> {
+  const vendors: string[] = [GO_VENDOR, ZEN_VENDOR, "copilot"];
+  if (vscode.workspace.getConfiguration(CONFIG_SECTION).get<boolean>(SETTING_AGENTS_WINDOW, true)) {
+    vendors.splice(2, 0, AGENT_GO_VENDOR, AGENT_ZEN_VENDOR);
+  }
+  const sections: string[] = [];
+
+  for (const vendor of vendors) {
+    const models = await vscode.lm.selectChatModels({ vendor });
+    sections.push(`## vendor: ${vendor}`, "", `models: ${String(models.length)}`, "");
+    for (const model of models) {
+      const internalModel = model as unknown as { configurationSchema?: unknown; detail?: unknown };
+      const schema = internalModel.configurationSchema;
+      sections.push(
+        `### ${model.name}`,
+        "",
+        `- id: \`${model.id}\``,
+        `- family: \`${model.family}\``,
+        `- version: \`${model.version}\``,
+        `- vendor: \`${model.vendor}\``,
+        `- detail: \`${typeof internalModel.detail === "string" ? internalModel.detail : ""}\``,
+        `- schema:`,
+        "```json",
+        JSON.stringify(schema ?? null, null, 2),
+        "```",
+        "",
+      );
+    }
+  }
+
+  const doc = await vscode.workspace.openTextDocument({
+    content: ["# OpenCode Model Picker Diagnostics", "", ...sections].join("\n"),
+    language: "markdown",
+  });
+  await vscode.window.showTextDocument(doc, vscode.ViewColumn.Beside);
+}
